Use upstream permissions for /usr/bin/fusermount
By default, /usr/bin/fusermount installs with the setuid bit set, and is
owned by root/root.  Setting 4750 permissions with 'trusted' group owner
is only useful if that group exists and certain UIDs are members of that
group, but that it is not currently the case.

This commit restores upstream permissions to /usr/bin/fusermount,
leaving the explicit %attr to make the setuid executable more visible.

Change-Id: I9871c9848b2b1bc5ad5f828911bf510bc72b62a2