Fix Launchpad bug #1087757 - fuse not building for aarch64
Tizen toolchain team in Samsung company is planning to build Tizen
for new arm 64 bit architecture - Aarch64. But package "fuse" cannot
be built for this architecture because of build errors.

The same problem appeared in open-source version of fuse during the
build on Launchpad for aarch64. It was fixed and the patch was made.

Bug: https://bugs.launchpad.net/linaro-oe/+bug/1087757
Change-Id: I8d9cdbccd0e788ffa959449371765cd171c3cdf4
Signed-off-by: Ilya Palachev <i.palachev@samsung.com>